Updated the README file
Updates to the readme file so that it's helpful for people learning the game, or anyone more interested in the project overall. The 'GithubPage' will be replaced with icon that will direct you to the respective persons github page, I will later add other ones if needed, like linkedIn or others.