Added the family menu
For now everything looks really bad - everything is using a placeholder. The timings of the event are set to fire too often, this is for testing purposes and will be changed later. Also for now, the event can kill the same person multiple times, must be changed later. Also, responding to the event doesn't cost anything yet, don't know how much it should cost - will be added later.